// src/net/bindings/types.ts
export const BeeState = __t.object("BeeState", {
  playerId: __t.u32(),
  cell: __t.u32(),
  x: __t.f32(),
  y: __t.f32(),
  z: __t.f32(),
  yaw: __t.i16(),
  pitch: __t.i16(),
  aimYaw: __t.i16(),
  aimPitch: __t.i16(),
  tick: __t.u32(),
  hp: __t.u16(),
  maxHp: __t.u16(),
  flags: __t.u8(),
});
export type BeeState = __Infer<typeof BeeState>;